+++ src/widgets/AttrScrollWidget.cpp
@@ -47,7 +47,7 @@ AttrScrollWidget::AttrScrollWidget(DocSheet *sheet, DWidget *parent)
     createLabel(gridLayout, 11, tr("Security"), fileInfo.bsafe);
     QString sPaperSize = QString("%1*%2").arg(fileInfo.iWidth).arg(fileInfo.iHeight);
     createLabel(gridLayout, 12, tr("Page size"), sPaperSize);
-    createLabel(gridLayout, 13, tr("File size"), Utils::getInputDataSize(fileInfo.size));
+    createLabel(gridLayout, 13, tr("File size"), Utils::getInputDataSize(static_cast<qint64>(fileInfo.size)));
 
     auto vLayout = new QVBoxLayout;
     vLayout->setContentsMargins(10, 10, 10, 10);
